backend-django-recommend/pcgg_recommend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.43
botocore==1.31.43
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
Django==4.2.5
django-storages==1.14
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
jmespath==1.0.1
joblib==1.3.2
mysqlclient==2.2.0
numpy==1.26.0
outcome==1.2.0
pandas==2.1.1
Pillow==10.0.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.6.2
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.12.0
selenium-async==0.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==1.26.16
wsproto==1.2.0
